xbmc/windowing/WinEventsSDL.cpp
Mac supports native full screen mode - it enlarges the application to full screen in a new screen. This is possibly more convenient than Kodi's full screen mode, which changes resolution, frame rate, etc.
However, the code has some check related to Linux causing the resize event to be ignored when switching to a window size equal to the screen size. This should not be done on Mac.